Refactor numeric types and clarify array layer logic
Replaced float with double for text positioning in ZenithView for improved precision. Updated Math.Clamp to use double bounds in FrameScheduler. Changed ArrayLayers multiplication order in TextureDesc for clarity.

+++ sources/Views/Zenith.NET.Views.Avalonia/ZenithView.cs
@@ -71,8 +71,8 @@ public class ZenithView : TemplatedControl, IZenithView
                                          fontSize * dpi,
                                          new SolidColorBrush(Colors.White) { Opacity = 0.98 });
 
-            float x = (float)(Bounds.Width - mainText.Width) / 2;
-            float y = (float)(Bounds.Height - mainText.Height) / 2;
+            double x = (Bounds.Width - mainText.Width) / 2.0;
+            double y = (Bounds.Height - mainText.Height) / 2.0;
 
             context.DrawText(shadowText, new(x + 1.0, y + 1.0));
             context.DrawText(mainText, new(x, y));
